+++ src/main/index.ts
@@ -1,5 +1,7 @@
 import { cleanEncodedInput } from './decode-utils.js';
+import { createKeyMapTransform } from './key-map-transform.js';
 import { createReferenceTransform } from './reference-transform.js';
+import { buildShareUrl, extractTokenFromUrl } from './url-utils.js';
 import {
 	AVAILABLE_CODECS,
 	DEFAULT_WEB_SHARE_CODECS,
@@ -23,6 +25,9 @@ const createClient = Object.assign(
 		defaultWebShareMaxLength: DEFAULT_WEB_SHARE_MAX_LENGTH,
 		defaultWebShareVersion: DEFAULT_WEB_SHARE_VERSION,
 		createReferenceTransform,
+		createKeyMapTransform,
+		buildShareUrl,
+		extractTokenFromUrl,
 		createEngine,
 		createNamedCodec,
 		createWebShareEngine
@@ -39,12 +44,15 @@ export type {
 	EngineCompressResult,
 	JsonUrlFactory,
 	JsonUrlValue,
+	KeyMapTransformOptions,
 	NamedCodecClient,
 	NamedCodecStats,
 	ReferenceTransformEntry,
 	ReferenceTransformOptions,
 	ShareCodecDefinition,
-	ShareTransform
+	ShareTransform,
+	UrlShareLocation,
+	UrlShareOptions
 } from './types.js';
 
 export default createClient;